requirements-lite.txt
# Lightweight Flask Backend for PythonAnywhere
# Total size: ~15MB (vs ~200MB with Prisma)

Flask>=3.0.0
Flask-Cors>=4.0.0
psycopg2-binary>=2.9.9
firebase-admin>=6.4.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
Pillow>=10.0.0
gunicorn>=21.2.0

